- fix some errors in the documentation; reported by Benno Schulenberg at:  - http://lists.gnu.org/archive/html/wdiff-bugs/2007-10/msg00000.html   - the error report:
In wdiff.texi the --statistics option for mdiff is described twice.
Also --help and --version occur twice; their second occurrences are
straight copies from the wdiff section, and mistakenly mention -h
and -V as synonyms -- these work for wdiff, but not for mdiff.

The Info doc further mistakenly mentions -c as synonym for
--ignore-case (should be -i, also for wdiff), and -a as synonym
for --autopager (should be -A, but only for mdiff).  It looks like
there may be several more errors like this.


(note: the --statistics option was completely removed from the mdiff
documentation because mdiff does not have such an option)