Split usage message translation strings.
Now we have separate strings for all option explanations, and also separate
strings for options with arguments.  This should allow us to add options
without requiring all strings to be re-translated.  It would also make it
easier for some strings to be present in several files while others are not,
and ensure consistent translations for strings with multiple occurrences.