Split Usage messages for easier translation.
Now there is a separate string for every line of Usage message output. That
way, new options might be added in the future without invalidating the
messages for existing options.  Furthermore it's now easier to share common
options between different executables.

The options themselves haven't been split, as arguments to some options
require translations, and as having them translated separately makes it
really difficult to align all options in a pleasing way, especially with
multi-byte encodings or double-width logograms.  So translators have both
the freedom and the responsibiluty to manage column alignment for their
language.

This change also turns a lot of "#if 0 ... #endif" into "/*** ... ***/" in
order to avoid unneccessary translations for these options.  This is in
response to an email by Benno Schulenberg:
http://lists.gnu.org/archive/html/wdiff-bugs/2008-06/msg00003.html