Fix a bunch of errors reported by gcc -Wall.
There were some missing header inclusions. For now I've included the
headers, but if someone reports portability issues, we might have to check
for them in configure and maybe provide fallbacks if they are unavailable.

There were also some variables which gcc reports might get used
uninitialized.  I initialized them to 0 or NULL, but I'm far from certain
that this is what was originally intended.  Might as well be that they are
an indication of more serious errors in those functions.

Most other changes are cosmetic things to prevent warnings, like added
braces and the likes.