Upgrade to gettext-0.17.
As gnulib-tool has to be run after gettextize/autopoint, because it
overwrites some files from gettext, we have to have the gettext
infrastructure in repository as long as we have gnulib there.  Probably best
this way, because gnulib officially only supports latest gettext.

Activated the gettext gnulib module as well.  This is in order to avoid
problems with gnulib overwriting m4 files that gettext depends on with a
version that unmodified gettext no longer can cope with.  In that situation,
I'd now expect gnulib to provide adjusted gettext files, resolving any such
issues.