Moved VC_LIST_EXCEPT exceptions from .x-* files to cfg.mk.
This is much cleaner, as it avoids having an extra revision-controlled file
for every rule with exceptions.  The old exception files are still supported
(and suitable if you happen to have long lists of such rules).  Therefore
the change is backwards-compatible and thus suitable for inclusion into
gnulib as well.