Adjust proper alignment of usage help for all commands.
To faciliate this, I've added a perl script which simply prints all msgstr
from a po file, in the order in which they are referenced in the source
files.  That gives all usage strings one after the other.  I also
implemented some primitive detection of overlong lines, and wrapped some
translations to fit common terminal widths.