po-file handling perl scripts: copyright, documentation, added to dist.
Added a copyright comment as well as some pod documentation to the
print-po.pl script and the msgitm.pm library it uses.  Adjusted Makefile.am
to include these two scripts, as they might be useful for translators, and
as translators from TP usually work on released tarballs not repository
checkouts.