Introduce --enable-experimental configure switch.
This switch controls installation of mdiff, unify and wdiff2. That way, an
unconfigured installation will only install wdiff, avoiding flooding the
maintainers with bug reports about applications they themselves (yes, this
includes me) don't really understand (yet).  On the other hand, the code
does get released to a somewhat wider audience if it is included in the next
wdiff release, so giving interested parties a way to toy with these apps
might be a good thing and help their development.

* configure.ac: generate substitutions from senable switch
* man/Makefile.am, src/Makefile.am: use substitutions to decide what to install
* man/Makefile.am, src/Makefile.am: auto-build binaries for help2man so that
  "make dist" without preceding "make" will succeed
* README: Mention new flag, cleaned up some obsoleted information