Reintroduce termcap configuration.
Search the following libs for a tputs function: termcap termlib ncursesw and
ncurses.  The first library to provide the tputs function will be linked in,
and termcap support enabled.  There is a --with-termcap configure option to
override this behaviour, e.g.  using --without-termcap or --with-termcap=foo
to select libfoo instead of an autodetected library.

The include statements have been adjusted to include term.h as well, as the
ncurses man page for tputs recommends.  Configure doesn't check if the
headers actually declare tputs; we simply assume the compiler will complain
if things don't work out as expected.