Pass sc_prohibit_empty_lines_at_EOF syntax check.
Except ABOUT-NLS from that check, as it is generated by gettext and
therefore not under our control.

Drop empty line from end of Makefile.am.  It was there because previous
gettextize invocations used to append stuff to that file, and we want such
additions separated.  But recent gettext behaves well, so let's be clean
here.