Pass sc_prohibit_always-defined_macros, introduce EXIT_ERROR.
The old use of EXIT_FAILURE with a nonstandard value of 2 was confusing. 
It's cleaner to pick a new name when using a new value.  So I replaced most
occurrences of EXIT_FAILURE with EXIT_ERROR (=2), in distinction to
EXIT_DIFFERENCE (=1).

sc_prohibit_always-defined_macros complains about the environ definition in
a comment in build-aux/warn-on-use.h which obviously doesn't make a lot of
sense.  We simply ignore build-aux altogether for this kind of check.