Deprecate --no-init-term, make --terminal a synonym to it.
Terminal initialization is only required when performing cursor movement,
which wdiff doesn't.  Therefore there should be no need to initialize the
terminal.  Furthermore, doing initialization can cause the screen to be
restored to its pre-initialization state upon deinitialization, thereby
clearing all wdiff output.

As a consequence, we never do terminal initialization.  This is what the old
--no-init-term option did.  Nevertheless we keep the shorter and more
intuitive --terminal (-t) as the documented and preferred option,
deprecating --no-init-term.  We still accept the option, but don't document
it in the online help, and use a dedicated section on compatibility in the
man page.