Include merges in ChangeLog.
This way, contributors to sidelines, the translations sideline in
particular, will be credited properly in the ChangeLog as well, not just in
the translation message files.  One (somewhat ugly) consequence is that many
messages appear twice, once for the sideline commit and once for the merge. 
Can't be helped just now.