Unify pipe stream code.
Now we have a single function to do the bulk of the pipe operation.  The two
frontends readpipe and writepipe are rather minimalistic.
The main motivation behind this change is using dup2 from gnulib instead of
the old and probably less portable code relying on the order in which file
descriptors are allocated by the operating system.