Disable some syntax checks to ensure we pass the rest.
As there is no portable alternative to strncpy, and the strncpy uses in this
application should be handling things careful enough, this seems safe.

Keeping 9 years of copyrights in a single line in the texi file doesn't
agree with my ideas of good style, so I'll rather disable the check.