[mdiff] Fix conflicting short option letter.
There were two cases of 't' so far, --expand-tabs and --tolerance. This
could lead to segmentation faults, since one of them expects an argument
while the other does not. The latter is now removed since it is not even
reported in the online help. Yet another -t option is removed from the
documentation.

The whole thing still looks like it would need a lot more work, with all the
FIXME lines it contains.